Feat : datatype 변경 및 service에서 test를 위한 코드 추가

+++ src/main/java/com/mynutrient/community/repository/PostsRepository.java
@@ -13,7 +13,7 @@ public interface PostsRepository {
 
     public long updatePost(Posts post);
 
-    public long deletePost(int postSeq);
+    public long deletePost(long postSeq);
 
     public List<Posts> selectPostsList();
